# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "FIN-03 Agricultural Climate-Aware Credit Risk API",
    "version": "1.0.0",
    "description": "Proposed contract skeleton. Replace generic request/response placeholders with detailed schemas before implementation."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/loan-applications": {
      "post": {
        "operationId": "createLoanApplication",
        "summary": "Create LoanApplication",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/loan-applications/{application_id}": {
      "get": {
        "operationId": "getLoanApplication",
        "summary": "Get LoanApplication",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "parameters": [
        {
          "name": "application_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique application_id identifier"
        }
      ]
    },
    "/loans/{loan_id}/repayment-schedules": {
      "post": {
        "operationId": "createRepaymentSchedule",
        "summary": "Create RepaymentSchedule",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "loan_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique loan_id identifier"
        }
      ]
    },
    "/loans/{loan_id}/repayment-events": {
      "post": {
        "operationId": "createRepaymentEvent",
        "summary": "Create RepaymentEvent",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "loan_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique loan_id identifier"
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}